+++ RELEASE_NOTES.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **Added** configuration profiles, to support multiple environments.
 - **Added** `api_host`, `api_proxy` and `api_user_agent` to config file.
 - **Added** `help` command for those who need more than `-h` and `--help`.
+- **Added** service endpoint to `check` command output.
 - **Changed** environment variables to use a `CLOUDSMITH_` prefix.
 - **Fixed** validation for `push` commands that require a distribution.
 - **Fixed** token endpoint failing because API key overrides login/password.